# cache struct type info, refactor fieldsByIndex